refactor: view/store -> view/perspectiveStore
helps distinguish it from other view stores
